+++ CHANGELOG.md
@@ -50,6 +50,8 @@ The deprecated model `simple_aperture` in `photon_collection` is removed.
   (See [!827](https://gitlab.com/esa/pyxel/-/merge_requests/827)).
 * Merge 'optical_psf_with_multiiwavelength' into 'optical_psf'.
   (See [!837](https://gitlab.com/esa/pyxel/-/merge_requests/837)).
+* Add capability to integrate photon in 'simple_conversion'.
+  (See [!848](https://gitlab.com/esa/pyxel/-/merge_requests/848)).
 
 ### Others
 * Bump pre-commit versions.